+++ src/app/user/user.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import {Component, computed, signal} from '@angular/core'; //signal umożliwia nam używanie sygnalizacji
 import {DUMMY_USERS} from "../dummy-users";
 const randomIndex = Math.floor(Math.random() * DUMMY_USERS.length); // bierze random indeks z tabeli DUMMY_USERS
 @Component({
@@ -9,15 +9,21 @@ const randomIndex = Math.floor(Math.random() * DUMMY_USERS.length); // bierze ra
   styleUrl: './user.component.css'
 })
 export class UserComponent {
-  public selectedUser = DUMMY_USERS[randomIndex] // nazwa pola na start bierze randomIndex element
+  // nazwa pola na start bierze randomIndex element
+  // signal powoduje, że każdy obiekt, który używa tej zmiennej zostanie
+  // powiadomiony, że wartość się zmieniła, zatem należy ją odświeżyć
+  public selectedUser = signal(DUMMY_USERS[randomIndex])
+  // Wywołujemy angularową funkcję computed , która jest powiązana z sygnalistą.
+  // Funkcja ta nam wybiera odświeżoną wartość z sygnalisty
+  public imagePath = computed(() => 'assets/users/' + this.selectedUser().avatar)
 
   // getter - metoda, którą możemy wywoływać z zewnątrz
-  get imagePath(){
-    return 'assets/users/' + this.selectedUser.avatar;
-  }
+  // get imagePath(){
+  //   return 'assets/users/' + this.selectedUser.avatar; // zanim użyliśmy signala
+  // }
 
   public getImagePath(){
-    return 'assets/users/' + this.selectedUser.avatar;
+    //   return 'assets/users/' + this.selectedUser.avatar; // zanim użyliśmy signala
   }
 
   /**
@@ -26,6 +32,7 @@ export class UserComponent {
   public onSelectUser(){
     console.log('Click');
     const randomIndex = Math.floor(Math.random() * DUMMY_USERS.length); // bierze random indeks z tabeli DUMMY_USERS
-    this.selectedUser = DUMMY_USERS[randomIndex];
+    // this.selectedUser = DUMMY_USERS[randomIndex]; zanim użyliśmy signala
+    this.selectedUser.set(DUMMY_USERS[randomIndex]);
   }
 }