+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 02.01.2025
+- Signals in user component
 - Improvement of user component
 - New user component
 - New header component